+++ tests/screenshot/apps/SandstoneComponents.js
@@ -28,6 +28,7 @@ import KeyGuide from './components/KeyGuide';
 import Layout from './components/Layout';
 import Marquee from './components/Marquee';
 import MediaOverlay from './components/MediaOverlay';
+import PageViews from './components/PageViews';
 import Panel from './components/Panel';
 import Panels from './components/Panels';
 import Picker from './components/Picker';
@@ -82,6 +83,7 @@ const components = {
 	Layout,
 	Marquee,
 	MediaOverlay,
+	PageViews,
 	Panel,
 	Panels,
 	Picker,