+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The following is a curated list of changes in the Enact sandstone module, newest
 - `sandstone/VirtualList` to show the focused item properly when switching from pointer mode to 5-way mode
 - `sandstone/VirtualList.VirtualGridList` to show the focused item fully when focus moves via 5-way keys inside a list
 
+### Changed
+
+- Popup components to show black background in high-contrast mode for better visibility
+
 ## [2.8.0] - 2024-02.07
 
 ### Fixed