+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
-## [2.9.0-alpha.1] - 2024-04-04
+## [2.9.0-alpha.1] - 2024-04-09
 
 ### Added